requirements.txt
pytest>=8.0.0,<9.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
